feat: Implementando os serviços de movies

+++ LarissaLavinia/ingresse/src/main/java/br/com/ingresse/repositories/IMovieRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ingresse.entities.Movie;
 
-public interface IMovieRepository extends JpaRepository<Movie, Integer>{
+public interface IMovieRepository extends JpaRepository<Movie, Long>{
 	
 }
